refactor: change directories to Pushie
Change from passpusher to pushie to match application name
