+++ app/build.gradle
@@ -5,7 +5,7 @@ apply plugin: "kotlin-android-extensions"
 android {
     compileSdkVersion 30
     defaultConfig {
-        applicationId "com.chesire.passpusher"
+        applicationId "com.chesire.pushie"
         minSdkVersion 23
         targetSdkVersion 30
         versionCode 1
