+++ settings.gradle
@@ -1,2 +1,2 @@
-rootProject.name='PassPusher'
+rootProject.name='Pushie'
 include ':app'
